requirements.txt
nltk==3.6.2
beautifulsoup4==4.9.3
pyspellchecker==0.6.2
googletrans==3.0.0
char-split==0.1.0.dev2
